refactor(api): move agent wrappers to own file
